Only delete if it's there

+++ models/ghe.rb
@@ -22,7 +22,7 @@ Model.new(:ghe, 'Description for ghe') do
   # Prep the backup - we need to get exports of all the data
   before do
     # Make sure we have a clean slate
-    FileUtils.rm_r(STAGING_PATH)
+    FileUtils.rm_r(STAGING_PATH) if File.exists?(STAGING_PATH)
     FileUtils.mkdir_p(STAGING_PATH)
 
     export_commands = [
